PHNX-2641: Fix for parallel run issues.
