+++ pom.xml
@@ -12,7 +12,7 @@
 
     <properties>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
-        <digyrunner.version>0.0.85</digyrunner.version>
+        <digyrunner.version>0.0.89</digyrunner.version>
     </properties>
 
     <repositories>
@@ -27,7 +27,7 @@
         <dependency>
             <groupId>org.seleniumhq.selenium</groupId>
             <artifactId>selenium-java</artifactId>
-            <version>4.8.3</version>
+            <version>4.12.0</version>
         </dependency>
 
         <dependency>
